requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
streamlit>=1.36
onnxruntime>=1.17
huggingface_hub>=0.24
pillow>=10.0
numpy>=1.26
pandas>=2.2
python-multipart>=0.0.9
